+++ src/lib/Scenes/Show/Components/ShowHeader/ShowHeader.tsx
@@ -37,7 +37,7 @@ export class ShowHeader extends React.Component<Props, State> {
     const {
       show: { partner },
     } = this.props
-    SwitchBoard.presentNavigationViewController(this, `/${partner.id}?entity=gallery`)
+    SwitchBoard.presentNavigationViewController(this, `${partner.href}?entity=gallery`)
   }
 
   handleFollowShow = () => {
@@ -209,6 +209,7 @@ export const ShowHeaderContainer = createFragmentContainer(
         ... on Partner {
           name
           id
+          href
         }
         ... on ExternalPartner {
           name