+++ CHANGELOG.md
@@ -17,11 +17,13 @@
 ### Master
 
 - Add ellipsis to partner and event names - ashkan
+- Fixes partner link on Fair Booth header - luc
 - Adds Sentry to the app - orta
 - Adds Sentry logger for users - orta
 - Fix for the count of artworks in a show - orta
 - Nearby show previews use a better placeholder - orta
 - Fixes to the fair icon - orta
+- Adds clickable links for maps - orta
 
 ### 1.8.4